loosen some req version constraints, and install gclogger separately (#172)
We were getting an error where the molfeat package wasn't downloading a
model properly. I found that by loosening some package version
constraints and installing the gclogger separately (because when
installing it with the other packages pip was taking a very long time
(>20m) trying to find compatible package versions). Doing this resolves
the error and the pip install cell runs more quickly.